requirements.txt
# Core dependencies
google-adk>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM support
langchain>=0.1.0
openai>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
